Avoid collision during IPsec cert renewal
The IPsec certificate renewal cronjob will be invoked at the same time
on all hosts when the certs are approaching expiry within 15 days.

But when hosts connected in IPsec transport mode do rekey/reauth
(a step in the certificate renewal) simultaneously, there could be
collisions between the two sides of the IPsec, racing to install new
SAs and policies, which can result in temporary disruption or failures
in reauthentication. While the SAs will eventually established, it may
take logner time and make mtc detect connectivity lost and trigger
host reboot.

This update introduced a per host unique delay in the renewal script,
so that the execution of the script doesn't overlap among hosts. This
way the collision will be avoided.

Test Plan:
PASS: Update CUTOFF_DAYS in the script (/usr/bin/ipsec-cert-renew.sh)
      to 90. Run the script to renew IPsec certificate. Verify the
      renewal is deplayed a random seconds within an hour, and the
      renewal is successful.
PASS: With CUTOFF_DAYS updated to 90 days, wait for the cronjob to run
      at mid night. Verify from cron.log that the script runs at
      different time on each host, and the renewal is successful on all
      hosts.

Closes-Bug: 2111362

Change-Id: I72c5cb80b1fcf8f0dd89b05fa30620021013f199
Signed-off-by: Andy Ning <andy.ning@windriver.com>