Remove SSSD "sudo rules" configuration
SSSD "sudo rules" configuration that supports sudo capabilities
for ldap users and groups has been replaced with pam-group
configuration. This commit removes "sudo rules" configuration
dormant code that has been kept to be readily available if
sudo rules would be eventually used in the future.
However, incorrect setting of the sssd service parameter
"ldap_search_base" for WAD domains affects the sudo parameter
"ldap_sudo_search_base", causing sssd failure.
To prevent such failures, even if there is a remote possibility
for them to occur, the "sudo rules" configuration should not pose
any risk and needs to be removed and this is done in this commit.
If in the future, there will be a need for sudo rules configuration,
it will be readded then.

Test Plan:
PASS: Verify the SSSD service is running correctly and there are
no alarms.
PASS: Configure SSSD for access to a WAD domain using service
paramter cli commands and verify SSSD connectivity to the WAD
server.
PASS: Verify that a WAD ldap user can ssh to the stx platform.
PASS: Verify that a local ldap user can ssh to the stx platform.

Closes-Bug: 2111466

Signed-off-by: Carmen Rata <carmen.rata@windriver.com>
Change-Id: Iaba61ff7e626593cf988f2d1d07778fc675f8c67