Bypass deferred manifests for SRIOV during enrollment
During enrollment, audit is disabled to avoid uncertainty with
manifests execution. However, it's necessary to run SRIOV manifest
if the user chooses to configure, as it's not possible to unlock the
system prior to this. This commit introduces a new parameter to
_config_apply_runtime_manifest to control if a call should be added
to the deferred manifest list or bypass it completely and execute
right away. This bypass method for SRIOV is only called during
enrollment prior to unlock, as the control flag is cleared after.

Test plan:
  - PASS: Enroll a subcloud configuring SRIOV and verify the operation
          completes successfully, the host is unlocked and dcmanager
          shows the subcloud as online after.

Closes-bug: 2106241

Change-Id: I146437802dce86433116ee627b583d0a7e1eec30
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>