Fix the application recovery process being aborted incorrectly.
An intermittent bug occurs when a user runs the abort command during
an app update process. The bug is triggered depending on the exact
timing of when the abort is executed.

If the abort command is executed before any modifications are made to
FluxCD, the recovery flow skips certain validations and completes
successfully. However, if that’s not the case, it reaches the
make_app_requests stage, where a specific section of the code checks
if the app has an abort status set. If it does, the routine aborts the
recovery process.

This commit adds a call to deregister_abort_status at the beginning of
the recovery function, fixing the issue of the recovery being aborted
incorrectly.

Test plan:
PASS: build-pkgs && build-image.
PASS: AIO-SX fresh install.
PASS: install and start the update of metrics-server app.
PASS: wait until the update start to apply the charts and then trigger
      the abort of metrics-server update operation.
PASS: checks if the system successfully recovery the app.

Closes-bug: 2111929

Change-Id: Ice79b65fb7f6e99df86ea2bda02d43d75a01d7b3
Signed-off-by: edias <edson.dias@windriver.com>