Avoid Creating Subcloud Routes on Worker Nodes
In a Distributed Cloud (DC) environment, if a non-controller node
(e.g., worker or storage) is added to the central cloud after subclouds
have already been deployed, the route for the subclouds may be
unintentionally added to the non-controller nodes.
This occurs because, when a worker node advertises its interfaces
(e.g., mgmt), the central controller mistakenly configures the
subcloud route through the worker’s management interface.
The controller assumes the interface belongs to a subcloud,
which is incorrect.

The root cause is that the system does not verify if host is controller or not.
The code must validate which node the mgmt interface actually belongs to before
applying subcloud routing logic.

Test done:

DC fresh install:
   2 controllers and 1 worker in the central cloud
   AIO-SX as subcloud
   Remove the worker node:
   	kubectl -n deployment delete host compute-0
   	or
   	system host-delete compute-0
   Add the worker node again with:
   	kubectl apply -f compute-0.yaml
   The worker node should be created without the route for subclouds

Closes-Bug: 2112143

Change-Id: I9feb7183e30d6de783f6fd5074df99e58579312b
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>