Fix Networktype checks in max_tx_rate/rx_rate flow
This commit fixes the existing checks enforced on rate-limiting
platform interfaces.
Earlier, an error is raised if an interface has no networktype
assigned. This was working fine in host-if-modify flow, as network
type is usually assigned by this time.

But the same checks are failing when an interface is newly created
via host-if-add or deployment manager with rate-limiting params.

Changes made -
Now we validate networktypeList of an interface only if the
networktype is already assigned.
In case networktypeList is empty, we will still allow for interface
addition/modification.

Affect -
Interfaces can get created/modified with max_tx_rate/rx_rate without
networktype assigned. Sysinv database will have max_tx_rate/rx_rate
values against these interfaces. But additional checks are already
in place in the file - sysinv/puppet/interface.py
We check the interfaces networktype while generating puppet hieradata

So, If a user who wants to apply rate-limiting, has to make sure
the right networktypes are assigned to the interface. This will
be highlighted in the feature documentation.

Test Plan -
PASSED: build-pkgs and designer iso deploy.
PASSED: A fresh interface can be added with max_tx_rate/rx_rate
        params without networkType via host-if-add flow.
PASSED: A fresh interface can be added with max_tx_rate/rx_rate
        params without networkType via the deployment manager flow.
PASSED: Hieradata is only created for interfaces with valid
        networktypes.

Change-Id: I5a9672fd412cf3b166b7346c7bfab76cf9ab1809
Signed-off-by: Aman Pandae <AmanPandae.Mothukuri@windriver.com>