Add semantic check of cgts-vg before unlocking
If a host is unlocked without any physical volumes associated
with cgts-vg, puppet will fail as it will attempt to remove
the VG that is in use. Below is an example of the error:

"""
 /Stage[main]/Platform::Lvm::Vg::Cgts_vg/Volume_group[cgts-vg]
 /physical_volumes: change from ['/dev/disk/by-path/pci-0000:b4
 :00.0-nvme-1-part4'] to [] failed: Execution of
 '/usr/sbin/vgremove cgts-vg --force' returned 5: Logical volume
 cgts-vg/root-lv contains a filesystem in use.
"""

To prevent this from happening, a semantic check has been
added to check for PVs associated with cgts-vg.

Additionally, improvements have also been made to the Agent:
- Do not send inventory message if there is an error
  getting PV or VG data.
- Do not notify the conductor if PV or VG data is empty.

Finally, it has also been modified to not allow forced
unlocking if the host has not yet been inventoried.

Test Plan:
 PASS: On Standard, manually delete the 'cgts-vg' from the
       'i_pv' table in the database and try to unlock the host.
 PASS: Fresh install on AIO-SX, AIO-DX, STD and Storage.

Closes-Bug: 2112588

Change-Id: I8d02cb909736fa724e8d013122dc7425226220bb
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>