Restore AdminEndpointRenew to cert-mon
When the certmon DC-specific code was migrated to dccertmon in [1],
the AdminEndpointRenew watcher, responsible for monitoring the
admin-ep secret in the subcloud and updating the
/etc/ssl/private/admin-ep-cert.pem file, was removed from cert-mon.

This causes the file to no longer be updated when the secret is
refreshed, leading to failures during subcloud management network
reconfiguration and subcloud rehoming in block [2].

This commit restores the AdminEndpointRenew watcher to cert-mon,
fixing this issue.

Test Plan:
PASS: Verify that AdminEndpointRenew runs in cert-mon service.
PASS: Manually delete the sc-adminep-certificate secret in the
      subcloud. Verify that the secret and the admin-ep-cert.pem
      file are correctly updated.
PASS: Manually delete the <sc>-adminep-ca-certificate secret in the
      SystemController. Verify that the secret and the
      admin-ep-cert.pem file are correctly updated in the subcloud.
PASS: Perform a subcloud rehoming. Verify that block [2] does not fail
      and that the admin-ep-cert.pem file is updated during the
      process.

[1]: https://review.opendev.org/c/starlingx/config/+/944987
[2]: https://opendev.org/starlingx/ansible-playbooks/src/branch/master/playbookconfig/src/playbooks/roles/common/update-sc-cert/tasks/main.yml#L59-L65

Closes-Bug: 2114865

Signed-off-by: Enzo Candotti <Enzo.Candotti@windriver.com>
Change-Id: Ifc11f3aca05af1efb7cd3ea251bda17ffaa4e67a