Create `force-deleting-with-app` state for istors
This changes introduces a new istor state `force-deleting-with-app` to
be used by the rook-ceph application.

When an istor is in this state, rook-ceph should skip data integrity
verifications to allow the removal in cases as the data held is already
lost due to a faulty disk.

This state is created when using rook-ceph and passing the option
`--force` to host-stor-delete:

  system host-stor-delete --force <uuid>

Or directly via istors endpoint:

  DELETE /istors/<uuid>?force=True

Test-Plan:
  PASS: Run `host-stor-delete --force <uuid>` and observe a configured
        istor state be transitioned to force-deleting-with-app
  PASS: Run `host-stor-delete <uuid>` and observe a configured istor
        be transitioned to deleting-with-app
  PASS: Request DELETE /istors/<uuid>  and ensure it behaves as
        force=False for retrocompatibility

Story: 2011372
Task: 52426

Change-Id: Ic0a6e67780549882a49474e43dd35c855171bcd7
Signed-off-by: Ítalo Vieira <italo.gomesvieira@windriver.com>