Prevent false positives when pushing to registry
Prevent push false positive results by inspecting container images after
the push API call to the local registry. This aims to robustify the push
process since the Docker API can return false positives when Docker
service is restarting. This allows the image download process to be
retried if pushing actually fails.

This issue was handled in a previous commit [1] by deferring, in the
conductor manager level, the execution of the puppet class
"platform::dockerdistribution::runtime". However, the agent manager,
which in fact runs puppet classes on conductor's behalf, does not have
any deferal logic. As a result, the conductor manager may allow the
class to run but the conditions may have changed when the agent manager
actually gets the chance to run it.

[1] https://review.opendev.org/c/starlingx/config/+/877724

Test Plan:
PASS: build-pkgs && build-image
PASS: AIO-SX fresh install
PASS: Inspect image successfully pushed to registry
      Confirm that the application apply process succeeds
PASS: Inspect image that was not pushed to registry
      Confirm that the image download process is retried

Partial-bug: 2115528

Change-Id: I69adf9b3d13a085b91b028365a32f5b56b1a2bdd
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>