Reject host-unlock when has in progress controllerfs
When the controllerfs ceph-float is added or deleted, and the unlock
action is triggered on the standby controller during the in-progress
state, the controllerfs end up in drbd_fs_update_error status, even if
the DRBD goes to UpToDate/UpToDate

To prevent this, a new check has been implemented to block unlocking the
standby controller if there is any controllerfs in in-progress status.
However, this check was only added for the non-forced unlock action, so
if you try to unlock the standby controller using the --force flag, it
will be possible.

Test Plan:
 - PASS: Fresh-install and apply the rook-ceph app with floating monitor
 - PASS: Lock/unlock the standby controller when the controllerfs is in
         available status and check if it is possible
 - PASS: Unlock the standby controller when the controllerfs is in
         in-progress (creating or deleting) status and check if it is
         not possible.
 - PASS: Unlock the standby controller when the controllerfs is in
         in-progress (creating or deleting) status using flag --force
         and check if it is possible.

Depends-On: https://review.opendev.org/c/starlingx/config/+/953887

Closes-Bug: 2115793

Change-Id: Id70187550a67dd8e239983867a03a4215b52dc26
Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>