Relocate SKIP_CEPH_OSD_WIPING Flag Removal
During the backup and restore process, when controller-1 is unlocked
before the workers perform PXE booting, it removes the flag related to
skipping the wipe of OSDs (SKIP_CEPH_OSD_WIPING), causing the workers
to have their OSDs wiped since the flag is no longer present.

To fix this, the removal of the flag now occurs at the end of the
B&R process, into complete_restore action.

Test Plan:

- PASS: Perform a fresh installation using a designer ISO
        and verify the system is set up correctly.

- PASS: Perform B&R on all systems, with and without
        the wipe_ceph_osds flag. (Ceph bare-metal and Rook-ceph)

- PASS: Perform a backup and restore test using both IPv4 and IPv6
        and verify the system recovers correctly. (Ceph bare-metal
        and Rook-ceph)

- PASS: Perform a backup and restore test with open, controller and
        dedicated model (Rook-ceph)

- PASS: During the backup and restore process, install/unlock the
        controllers first and only then perform the PXE booting of
        the workers.

Closes-Bug: 2115925

Change-Id: I3e53b608e2eb961247ebc094658083f5e4eb6274
Signed-off-by: Gabriel Przybysz Gonçalves Júnior <gabriel.przybyszgoncalvesjunior@windriver.com>