Use legacy port for k8s health check during upgrade
During the upgrade to sxt 11, the port for kube-apiserver will be
changing from 6443 to 16443. This change is only completed after
the upgrade activation, which makes the health checks for k8s
made using sysinv fail.

This review introduces code to handle the port used in the health
checks, using the legacy port when the modification is not
completed.

Test plan:
PASS: Deploy AIO-SX.
PASS: Upgrade AIO-SX, AIO-DX from stx 10 to stx 11.

Story: 2011399
Task: 52466

Change-Id: Ie8a987e81f57e45d3af28ab60a1fb03d67824b27
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>