Change rook-ceph check in restore-complete
With the changes made to rook-ceph's BnR, it is necessary to
change the check whether rook-ceph recovery is still in progress
or not. This is set according to the status defined in the
rook-ceph-recovery configmap, as follows:

recovery-completed: It has already been completed and will pass.
recovery-failed: An error occurred and the restore-complete
                  cannot be performed. The reason for the failure
                  is displayed.
any other status: Recovery is not yet complete and needs to wait.

Test Plan:
 - PASS: On AIO-SX, run 'system restore-complete' after
         controller-0 is unlocked.
 - PASS: On AIO-DX, force an error in rook-ceph recover
         and run 'system restore-complete'.
 - PASS: On Standard, while the 'rook-ceph-operator' job is
         running on controller-1, run 'system restore-complete'.

Story: 2011372
Task: 52517

Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/953340

Change-Id: Iffbbdba0bedeb9a4d4864fdc88b5dac5ec0691f7
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>