Kube-apiserver port change upgrade rollback
During the upgrade to stx 11, the port for kube-apiserver will be
changing from 6443 to 16443. This change needs to be reverted in
case the upgrade script that applies it was already executed
before the abort/rollback.

This review introduces a RPC method that will be called during the
upgrade activation to trigger required puppet manifest in the
controller hosts, to create a flag that signals that a rollback
is happening.

Test plan:
PASS: Upgrade activation - AIO-SX and DX (stx10->11).
      Verified no stuck 250.001 alarms after the script execution.
PASS: Rollback after upgrade activation - AIO-SX and DX (stx11->10).

Story: 2011399
Task: 52488

Change-Id: If71a4266c178786b3fab7dd568cff88bd7b479ff
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>