Do not deprecate mgmt, admin and storage floating address in AIO-SX
With the implementation of story "Platform Network Address Reduction
 - AIO-SX" (2011191) only one address is configured in Linux per
network, the floating one. This will affect all platform networks and
there is no need to mark the floating one as deprecated.

The address deprecation is used to prevent the usage of the address
as the source IP in any communications, up until now, in AIO-SX, the
management and admin networks configured 2 addresses (floating and
controller-0), so it was necessary mark the floating one as
deprecated. This is still necessary in DX as the floating address can
move between controller-0 and controller-1.

This change prevents adding the address to the hieradata variable:
"platform::network::addresses::address_config". This prevents the
deprecation in simplex systems

Test Plan
=========
[PASS] Instal AIO-DX in IPv6 and dual-stack (IPv6 as prim and
       secondary) and check that floating addresses are set as
       deprecated for mgmt, admin and storage
[PASS] Instal AIO-SX in IPv6 and dual-stack (IPv6 as prim and
       secondary) and check that floating addresses are NOT set as
       deprecated for mgmt, admin and storage

Story: 2011191
Task: 52520


Change-Id: I07887039303bd404a1ccaa8f23d4f5b97a00af0b
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>