Handle access forbidden error message
In controllers with enforce policy, when the access is denied it
raises a forbidden exception since the error message is not set,
the client is showing 'None' as response in the cli. This fix
addresses this issue by setting a custom error message also
the exception is handled at the client.

Test Cases:
PASS: Execute the following as reader role and expect
       "The requested action is not authorized"
       system host-update controller-0 location="{'city':'delhi'}"
       system service-parameter-modify platform client cli_confirmations="enabled"
       system service-parameter-apply platform

Story: 2011348
Task: 52525

Change-Id: Ia347bff72678cc3ccfcd91cbf5915afea50c0ac7
Signed-off-by: amantri <ayyappa.mantri@windriver.com>