K8s upgrade refactor: unit tests for image download
The Kubernetes upgrade currently relies on several external scripts
that are implemented using Ansible and Puppet depending on the
specific step being executed.  In order to reduce the overall
execution time and to cleanup the current implementation, the
dependency on external scripts is eliminated and implementation
is moved to sysinv. Equivalent RPC APIs are added to sysinv-agent
to replace the code in puppet manifests.

Unit tests are added in this change for refactored image download
code that was added in the following change:
https://review.opendev.org/c/starlingx/config/+/952008

This change also makes below improvements in the code:
1. Method docstring
2. Ignore failures gettting existing crictl image list that is
   looked up before pulling an image to crictl. Image pull can be
   continued regardless of the exisitng image list is obtained or
   not.

Testing:
PASS: K8s upgrade from 1.29->1.30 successful in AIO-DX lab.
      System healthy, No alarms, K8s cluster healthy after the
      upgrade.
PASS: All unit tests passed.

Story: 2011477
Task: 52303

Change-Id: I84eb4996994ee23d3424015b578f29915c284e7c
Signed-off-by: Kaustubh Dhokte <kaustubh.dhokte@windriver.com>