K8s upgrade refactor: unit tests for networking upgrade
The Kubernetes upgrade currently relies on several external scripts
that are implemented using Ansible and Puppet depending on the
specific step being executed.  In order to reduce the overall
execution time and to cleanup the current implementation, the
dependency on external scripts is eliminated and implementation
is moved to sysinv. Equivalent RPC APIs are added to sysinv-agent
to replace the code in puppet manifests.

Unit tests are added in this change for refactored networking upgrade
code that was added in the following change:
https://review.opendev.org/c/starlingx/config/+/952009

This change also makes below improvements in the code:
1. Robustify kubectl_apply method parameter type checking
2. Cleanup of upgrade overrides file in case of both success
   and failure cases of networking upgrade.

Testing:
PASS: K8s upgrade from 1.29->1.30 successful in IPv4 lab. (AIO-DX)
      System healthy, No alarms, K8s cluster healthy after the
      upgrade.
PASS: All unit tests passed.

Story: 2011477
Task: 52304

Change-Id: Ie273753303a8a195c928a1a023aaf03cfa8b6e53
Signed-off-by: Kaustubh Dhokte <kaustubh.dhokte@windriver.com>