K8s upgrade refactor: unit tests for storage upgrade
The Kubernetes upgrade currently relies on several external scripts
that are implemented using Ansible and Puppet depending on the
specific step being executed.  In order to reduce the overall
execution time and to cleanup the current implementation, the
dependency on external scripts is eliminated and implementation
is moved to sysinv. Equivalent RPC APIs are added to sysinv-agent
to replace the code in puppet manifests.

Unit tests are added in this change for refactored storage upgrade
code that was added in the following change:
https://review.opendev.org/c/starlingx/config/+/952010

This change also makes improvements in the code such as:
1. Method docstring
2. Variable name typos

Testing:
PASS: K8s upgrade from 1.29->1.30 successful in AIO-DX lab.
      System healthy, No alarms, K8s cluster healthy after the
      upgrade.
PASS: All unit tests passed.

Story: 2011477
Task: 52305

Change-Id: Id4985141d75ea5e3fe7cd190248fe3805c05a83c
Signed-off-by: Kaustubh Dhokte <kaustubh.dhokte@windriver.com>