K8s upgrade refactor: unit tests for kubelet upgrade
The Kubernetes upgrade currently relies on several external scripts
that are implemented using Ansible and Puppet depending on the
specific step being executed.  In order to reduce the overall
execution time and to cleanup the current implementation, the
dependency on external scripts is eliminated and implementation
is moved to sysinv. Equivalent RPC APIs are added to sysinv-agent
to replace the code in puppet manifests.

Unit tests are added in this change for refactored kubelet upgrade
code that was added in the following change:
https://review.opendev.org/c/starlingx/config/+/952732

This change also makes below improvements in the code:
1. Correct method call that reports error in case of unexpected
   exception in kubelet upgrade Sysinv-agent RPC API call from
   sysinv-conductor.
2. Correct method docstring.
3. Correct code comments.

Testing:
PASS: All unit tests passed.

Story: 2011477
Task: 52307

Change-Id: Ic781a830e3ce1acd8cd36dc30d94b01591562d5a
Signed-off-by: Kaustubh Dhokte <kaustubh.dhokte@windriver.com>