Fix children deletion when host is deleted
When a host is deleted, the associated data (interfaces, ports, lvg,
pv) are not deleted but the foreign key field was set to null.
This behavior was introduced when SQLAlchemy was up-versioned to 1.4.
To fix this, we need to define cascade on the backref of the parent in
relationship on the child class.

Test plan:
PASS - Host delete and re-add
PASS - Delete and re-add interface

Closes-bug: 2118773

Change-Id: Ia1e53abedb95698db16cd8a07f682532aef08eb0
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>