Provide the host management address to MTCE api
The field mgmt_ip was deprecated in the ihost table, the value needs
to be obtained from the address table based on the hostname. Logic was
added to consider the following cases
a) management reconfiguration, in AIO-SX the management could have
been already changed at the database, but not applied on the OS. The
old management addresses are still configured, in this case the
current OS configuration needs to be sent to MTCE, the updated mgmt
addresses will be sent after unlock
b) SX-to-DX migration: if in the middle of the migration, retrieve
the floating address even if the hostname is controller-0, as SX
doesn't have unit addresses, only floating one.

Test Plan
=========

[PASS] install an STD and monitor the host mgmt IP sent to MTCE in
       the case of forced lock
[PASS] install AIO-SX and change the management address, observe the
       correct address is sent to MTCE in a forced lock situation
[PASS] install an AIO-DX, remove controller-1, reinstall as a clone
       of the original node.
[PASS] Install AIO-SX and migrate to DX. checking that MTCE receives
       the correct address.

Closes-Bug: 2120882

Change-Id: I86ccdfcf7b4a8b5120c4c4e88ba75552c45f98a8
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>