Fix Helm release removal on application updates
This commit fixes a bug where Helm releases from the N version of an
app are not removed if they do not exist in the N+1 version.

Prior to this fix, an exception was being raised because deprecated
charts from N version are not enabled and are removed from the database
during the app removal process. This change accounts for that by
retrieving all N version application charts that are candidates for
removal, regardless of whether they are enabled on the database or not.
Once releases are uninstalled, the unused corresponding charts are
removed from the repository.

In addition, the metadata docstring was updated to include the missing
class setting.

Test Plan:
PASS: build-pkgs && build-image
PASS: Update istio app to a version that does not contain the
      istio-operator chart.
      Confirm that the Helm release and the corresponding chart were
      removed after the update.
PASS: Downgrade istio app to a version that does not contain the
      istio-istio-base, istio-istio-cni, istio-istio-egressgateway,
      istio-istio-ingressgateway and istio-istio-pilot charts.
      Confirm that Helm releases and corresponding charts were removed
      after the downgrade.

Closes-bug: 2119115

Change-Id: Ica9384aa0c3fca53b71d12e6fb4bcc7419fa729a
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>