Fix sysinv application-isolated cores validation
There is a restriction in sysinv api that prevents configuring both
vSwitch and application-isolated cores on the same processor node. In
systems using the vSwitch type OVS-DPDK, this configuration is common
and should not be restricted. However, currently, it is only possible
to configure vSwitch and application-isolated cores on different
processor nodes.

The deprecated validation that fails and triggers the error checks for
the openstack-compute label when application-isolated cores are
present, a condition that is not applicable considering that this
restriction should not exist and the labels have not yet been
configured at that stage. To solve this issue, the validation check
was removed.

TEST PLAN:

PASS - build-pkgs -c -p sysinv
PASS - build-pkgs --all
PASS - build-image
PASS - AIO-SX fresh install deploying vswitch and application-isolated
       cores on the same processor node
PASS - AIO-DX fresh install deploying vswitch and application-isolated
       cores on the same processor node

Closes-Bug: 2119339

Change-Id: I60966d398c0931cc538702745abe54874f522b28
Signed-off-by: Murillo Arantes <murillo.arantes@windriver.com>