SR-IOV configuration updates on interface deletion
This change ensures that SR-IOV configuration is updated properly when
an interface is deleted and is associated with a data network.
Previously, the SR-IOV update was only triggered when a user removed the
interface data network. If the user deleted the interface directly, the
SR-IOV update logic was not executed, leaving the SR-IOV configuration
out of sync. This fixes the issue by triggering the SR-IOV update logic
when an interface is deleted and still associated with a data network.

Test Plan
=========

[PASS] Configure a data network interface by associating an SR-IOV
interface with a data network. Then, delete the interface and verify
that the SR-IOV resources are properly updated, both in the
'/etc/pcidp/config.json' file and via 'kubectl describe nodes'.

[PASS] Configure a data network interface by associating an SR-IOV
interface with a data network. Then, delete the interface datanetwork
and verify that the SR-IOV resources are properly updated, both in the
'/etc/pcidp/config.json' file and via 'kubectl describe nodes'.

Closes-Bug: 2119672

Change-Id: I1f57d5a9e0df42a4157c9e7ff5da4d63c0ff8230
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>