Circular dependency scenario taking a long time to fail
This change fixes the long delay in failing to install apps with
circular dependencies with a dependency root greater than 1. For
example:

app-x depends on app-y
app-y depends on app-z
app-z depends on app-x (Closes the dependency loop)

This scenario should fail, and it was happening, but it was taking
the 30-minute timeout that app-z waited for its dependency (app-x)
to install.

Apps with circular dependencies are now stored in the apps_metadata
attribute of the ConductorManager class. The solution here is to
simply check if the app being installed is present in this list. If
it is, it will return an immediate failure with an appropriate reason.

Test Plan:
PASS: upload, apply, remove, delete metrics-server app
PASS: upload, apply, remove, delete vault app
PASS: upload, apply, remove, delete intel-device-plugin app
PASS: Install apps with dependencies (without circular dependencies)
      and have them installed successfully.

Testing the fix in the apply scenario:
PASS: Install  an app with a direct circular dependency and
      observe the immediate app failure. For example:
         app-x depends on app-y
         app-y depends on app-x (Closes the dependency loop)
PASS: Install  an app with an indirect circular dependency and observe
      the immediate app failure. For exemple:
         app-x depends on app-y
         app-y depends on app-z
         app-z depends on app-x (Closes the dependency loop)

Testing the fix in the update scenario:
PASS: Update an application where your new version has a direct
      circular dependency and observe the immediate app failure.
       For exemple:
         app-x-1 will be update to app-x-2 that depends on app-y
         app-y depends on app-x (Closes the dependency loop)
PASS: Update an appwhere your new version has indirect circular
      dependency and observe the immediate app failure. For exemple:
         app-x-1 will be update to app-x-2 that depends on app-y
         app-y depends on app-z
         app-z depends on app-x (Closes the dependency loop)

Closes-bug: 2120336

Change-Id: Ibf8889ed104ba973b7e0de5392b8a93800efa4c2
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>