Prevent lifecyle from running multiple times
This commit reduces the amount of calls to the status lifecycle hook
executed during application operations.

Prior to this change, the status lifecycle was executed multiple times
for the same Helm release, leading to unnecessary overhead and several
repeated log messages.

That behavior was modified to execute the hook when the Helm release is
reported as ready, thus running it just once per release. In addition,
the release name is now being passed as an extra information to the
lifecycle hook, which can be useful for application developers in the
future.

This allows the correct execution of the lifecycle implemented to assert
cert-manager readiness:
https://review.opendev.org/c/starlingx/cert-manager-armada-app/+/956284

Test plan:
PASS: build-pkgs && build-image
PASS: Platform fresh install
PASS: upload/apply/update/remove/delete cert-manager
PASS: upload/apply/update/remove/delete rook-ceph

Partial-bug: 2119681

Change-Id: I32d10ed67273ba3b23f045ed8976106ac0ddf1a4
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>