K8s upgrade refactor: unit tests for control-plane upgrade
The Kubernetes upgrade currently relies on several external scripts
that are implemented using Ansible and Puppet depending on the
specific step being executed.  In order to reduce the overall
execution time and to cleanup the current implementation, the
dependency on external scripts is eliminated and implementation
is moved to sysinv. Equivalent RPC APIs are added to sysinv-agent
to replace the code in puppet manifests.

Unit tests are added in this change for refactored control-plane
upgrade code that was added in the following change:
https://review.opendev.org/c/starlingx/config/+/954006
Along with the above change, this also covers unit tests for the k8s
upgrade skew policy code added in the conductor/manager.py in this
change: https://review.opendev.org/c/starlingx/config/+/954138

This change also makes below improvements in the code:
1. Robustify _update_symlink input parameters value check
2. Seperate common code between kube_upgrade_apply and
   kube_upgrade_node methods and add it to common/utils.py as a
   generic and reusable code.

Testing:
PASS: AIO-DX K8s 1.29->1.30 upgrade successful.
PASS: All unit tests passed.

Story: 2011477
Task: 52308

Change-Id: I16a5dccc1fb831380c32480532a36ebcb160b41c
Signed-off-by: Kaustubh Dhokte <kaustubh.dhokte@windriver.com>