K8s external requests with client certs (admin)
A parent change for stx-puppet will allow external clients that
provide client certificates with credentials for the user
kubernetes-admin to connect with kube-apiserver. This is made by
having a separate backend where haproxy itself authenticates with
admin credentials to kube-apiserver, while verifying the external
client certificate in the frontend before redirecting to it.

As such, the client certificate used by haproxy to authenticate into
kube-apiserver needs to be refreshed when the kube-cert-rotation.sh
script renews the client certificate in the admin.conf kubeconfig,
and haproxy needs to be restarted to pick the refreshed bundle.

Also, this change fixes the API /kube_clusters to return the
haproxy port as well as the system-local-ca's RCA, for the default
cluster. This API provides the endpoint and admin credentials to
access kube-apiserver using the OAM.

Test plan:
PASS: Deploy AIO-SX.
      Tested external requests to kube-apiserver with client
      certificates and kubernetes-admin credentials.
      Tested external requests to kube-apiserver with client
      certificates and non kubernetes-admin credentials (negative
      test).
      Tested external requests to kube-apiserver with token.
      Run manually the script kube-cert-rotation.sh, verified that the
      bundle file with the client certificate for haproxy was updated,
      and haproxy is working as expected.
      Run 'system kube-cluster-show kubernetes', verified that the url
      returned in 'cluster_api_endpoint' has haproxy port (6443) and
      the certificate returned in 'cluster_ca_cert' is the RCA cert
      for system-local-ca.

Story: 2011399
Task: 52647

Depends-on: https://review.opendev.org/c/starlingx/stx-puppet/+/956799

Change-Id: I305c0179b4bf1870d1a60cfb321ea56be956df13
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>