Allow controller fs resize to proceed when upgrade activate or rollback
During upgrade, if filesystem size is modified, the state was not
getting updated to "Available". The state update should happen in
sysinv audit with "_controller_config_active_apply" function. A
check in audit for upgrade in progress prevents it to run and blocks
state change.

This commit runs the audit for certain deploy states, so that drbd
state change is possible after a size update.

Test Plan:
    Pass: Verify state change to Available on running
          "system controllerfs-modify platform" command
          when upgrade is in progress
    Pass: Verify state change after activation scripts are run
          on AIO-DX
          * activate-done and rollback done, failed
          * activate-failed with retries
    Pass: Verify firewall etc are not run (unnecessarily) after
          activate

Closes-Bug: 2120362
Change-Id: Ie484af32855552b317f98e448dd834461f8d41d3
Signed-off-by: sshathee <shunmugam.shatheesh@windriver.com>