K8s upgrade refactor: handle sysinv-agent restart
The Kubernetes upgrade previously relied on several external scripts
that are implemented using Ansible and Puppet depending on the
specific step being executed.  In order to reduce the overall
execution time and to cleanup the current implementation, the
dependency on external scripts is eliminated and implementation
is moved to sysinv. Equivalent RPC APIs are added to sysinv-agent
to replace the code in puppet manifests for following operations:

1. Download control plane images on the inactive controller
2. Control-plane upgrade
3. Kubelet upgrade
4. Kubernetes upgrade abort on a particular host.

In case sysinv-agent process gets killed or fail-restarted
unexpectedly while sysinv-agent part of any of above operations was
in progress, kubernetes upgrade is stuck forever in the state it is
in.

This change adds a functionality to handle this situation.
The kubernetes upgrade method name and argument details are saved
in a file when the operation is started and cleaned-up when done.
In case of unexpected service restart, the file is accessed if
present and the procedure is re-executed. If the file is not found,
it is assumed that upgrade was not in progress and no action is
taken. If for some reason, unfinished upgrade details are not found
in the file, it is reported to the conductor and kubernetes upgrade
is updated to mark the ongoing stage as failed.

Test Plan:
AIO-DX:
PASS: Start k8s upgrade 1.29->1.30.
      - Gracefully restart sysinv-agent on inactive controller while
        images being downloaded on the inactive controller.
        Image download is re-run after service restart.
      - Gracefully restart sysinv-agent on inactive controller while:
        1. images being downloaded on the ACTIVE controller.
        2. K8s networking is being upgraded
        3. K8s storage is being upgraded
        Observe that the operations are not impacted.
      - Gracefully restart sysinv-agent on the controller on which
	control-plane is being upgraded. Control-plane upgrade is re-run
        after service restart.
      - Gracefully restart sysinv-agent on the host on which kubelet
        is being upgraded. Kubelet upgrade is re-run after
        service restart.

PASS: Start k8s upgrade 1.29->1.30.
      - Corrupt the kubernetes procedure details file and gracefully
        restart sysinv-agent on the inactive controller while
        images being downloaded on the inactive controller.
        Kubernetes upgrade is updated to "download-images-failed"
      - Corrupt the kubernetes procedure details file and gracefully
        restart sysinv-agent on the controller on which control-plane
        is being upgraded. Kubernetes upgrade is updated to
        "upgrading-first-master-failed" or
        "upgrading-second-master-failed" depending on which operation
        was in progress. Host upgrade status is updated to
        "upgrading-control-plane-failed"
      - Corrupt the kubernetes procedure details file and gracefully
        restart sysinv-agent on the host on which kubelet is being
        upgraded. Host upgrade status is updated to
        "upgrading-kubelet-failed"

AIO-SX:
PASS: Start k8s upgrade 1.29->1.30. Abort it after network upgrade.
      - Gracefully restart sysinv-agent while abort was in progress.
        Abort is re-run after the service starts.
      - Corrupt the kubernetes procedure details file and gracefully
        restart sysinv-agent while abort was in progress, kubernetes
        upgrade status is updated to upgrade-abort-failed.

PASS: Restart sysinv-agent when no k8s upgrade is running.
      No action is taken.

Closes-bug: 2120981

Change-Id: I9a6fdfe75dbeeb71ec220f7994c9f010b05df023
Signed-off-by: Kaustubh Dhokte <kaustubh.dhokte@windriver.com>