Support larger size key in system-local-ca for IPsec
Currently multi node system bootstrap will fail on IPsec configuration
task if system-local-ca is overridden with certificates that have larger
keys (eg, 8192 bit).

This is because it takes longer time for cert-manager to approve the
certificaterequest and make the IPsec certificate available when the key
is bigger.

This change fixed this by adding retries to retrieve the certificate if
it's not available in the first attempt.

Test Plan:
PASS: Bootstrap a DX system with overrides to system-local-ca that has
      8192 bit keys. Verify the bootstrap is successful, IPsec
      configuration file, certificate and key are generated, and IPsec
      connections are loaded.

Closes-Bug: 2120681

Change-Id: Iefeaa9cfbcda3719081e785161959d2724272f02
Signed-off-by: Andy Ning <andy.ning@windriver.com>