Allows multiple user credentials in same session
This change adds the user identification to the key name used to manage
the credentials stored in the keyring, allowing multiple user
credentials in the same terminal session.

TEST PLAN:
PASS: Build package and ISO.
PASS: Without source and passing the user credentials on the command
      line, check if the user has the expected permissions.
PASS: Source and verify that the user permissions are in place.
PASS: Switch users and verify that the correct credentials are being
      used.

Closes-Bug: 2120721

Change-Id: Iaf2d61bf02c7075ef22bdb3ac95aa9137f51e192
Signed-off-by: Reynaldo P Gomes <reynaldo.patronegomes@windriver.com>