Fix adding ceph backend after provisioned/locked
When Ceph is configured with the controller locked only
after the first unlock, puppet fails because the
ceph-mon-lv size is zero.

This happens because with the change made in [1],
'platform::ceph::params::mon_lv_size' is only set in
hieradata when there is already a mon configured on
the host.

To fix this, changes were made so that the backend is
configured on unlocking, and the backend and ceph-mon
states reflect the correct state; that is, it is only
updated when it is actually configured.

[1]: https://review.opendev.org/c/starlingx/config/+/926020

Test Plan:
- PASS: Fresh install on AIO-SX/AIO-DX/STD without backend defined
- PASS: Lock controller-0
- PASS: Add ceph backend and OSDs
- PASS: Unlock controller-0
- PASS: Check that ceph has been configured and
        that there are no errors in puppet.

Closes-Bug: 2091578

Change-Id: I236a3794ce9e08dfebb816e9f4a4cbdc19ab7e69
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>